falla una consulta de alta en ventana altas bajas, no se como insertar el offline

+++ src/util/DataBaseControler.java
@@ -48,14 +48,27 @@ public class DataBaseControler {
 	
 	//___________________Altas/bajas/loggin_____________________________//
 	
-	public void  alta(String nick,String pass) throws SQLException{
+	public String alta(String nick,String pass) throws SQLException{
+		String respuesta;
+		if (!nick.equals("")&& !pass.equals("")){
 		Statement st = conn.createStatement();
-		ResultSet rs2 = st.executeQuery("insert into User values('"+nick+"','"+pass+"';");
-		
+		ResultSet rs2 = st.executeQuery("insert into User values('"+nick+"','"+pass+"',offline;");
+		respuesta = "211.OK Usuario insertado correctamente";
+		}else{
+			respuesta = "411.ERR Faltan datos";
+		}
+		return respuesta;
 	}
-	public void  baja(String nick,String pass) throws SQLException{
+	public String  baja(String nick,String pass) throws SQLException{
+		String respuesta;
+		if (!nick.equals("")&& !pass.equals("")){
 		Statement st = conn.createStatement();
 		ResultSet rs2 = st.executeQuery("delete from User where nick ='"+nick+"' and where contrasena ='"+pass+"';");
+		respuesta = "210.OK Usuario eliminado correctamente";
+		}else{
+			respuesta = "410.ERR Faltan datos";
+		}
+		return respuesta;
 	}
 
     public String consultaUsuario(String nombre) throws SQLException {
